// src/app/models/animal.model.ts
export interface Animal {
  id: number;
  name: string;
  age: string;
  race: string;
  gender: string;
  species: string;
  color: string;
  size: string;
  weight: number;
  height: number;
  physicalDescription: string;
  vaccinated: boolean;
  castrated: boolean;
  microchipped: boolean;
  observation: string;
  status:string;
  createAt: Date;
}